+++ changelog.md
@@ -3,6 +3,7 @@
 * Fix rare crash when exiting a level
 * Fix some crashes in the level browser on android
 * Make the "outdated client" message more straightforward
+* Add player counts to tower levels (**requires** Node IDs mod to be installed first)
 * Fix progressbar icons having wrong glow when the Colored Icon Kit mod is installed
 * Try to fix progressing in levels with startpos (again)
 * Try to fix crashes when using a Level Edit hack